Move credential issuance auth logic outside entities
