+++ src/testUtils/db.ts
@@ -10,7 +10,7 @@ import { Member } from '../entities/members/Member.model.js';
 import { MemberPublicKey } from '../entities/memberKeys/MemberPublicKey.model.js';
 import { MemberKeyImportToken } from '../entities/memberKeyImports/MemberKeyImportToken.model.js';
 import { MemberBundleRequestModel } from '../entities/memberKeys/MemberBundleRequest.model.js';
-import { CachedJwks } from '../entities/memberSignatures/exchanges/CachedJwks.model.js';
+import { CachedJwks } from '../credentialIssuanceAuth/CachedJwks.model.js';
 
 const MODEL_SCHEMAS = Object.values([
   CachedJwks,
